Fixed crash for PAStats IRC
PAStats IRC window crashed if it received a username that was not a valid irc nickname (due to character range limitations).
This fix adds a check whether it is valid and uses a default proposal if it isn`t.

The real problem is within iris, the client used by webchat.esper.net .